requirements.txt
agentpy==0.1.5
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dill==0.3.8
Flask==3.0.3
fonttools==4.53.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
multiprocess==0.70.16
networkx==3.3
numpy==2.1.0
owlready2==0.46
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
SALib==1.5.1
scipy==1.14.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
